Remove relative ordering of head display
When all heads are on the side, or some are disabled, it can get just
more confusing. Fixing it would be difficult and I'm not sure if it
would be worth it. Also, I feel lazy lately.

With this change, the heads display order is fixed, left->right,
top->bottom.

Signed-off-by: Petr Horacek <hrck@protonmail.com>