Set the shortest save interval to 1 second
With this change, save requests are accumulated over 1 second before
storing.

This makes sure that even under lot of load and rapid changes of
storable options, the CPU won't be overwhelmed by saving.

This is addressing an issue where mapping CV input to SPEED while
the current tempo is tapped in would produce 2 saves one after the
other, stressing the module and lowering the idle time momentarily
below 7 percent.

Signed-off-by: Petr Horacek <petr@zlosynth.com>